requirements.txt
python-dateutil
requests
sco-datastore
sco-engine
